+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Restores the ability to specify a Hosting `site` (configured in `firebase.json`) as a deploy `--only` target. (#2731)
+- Introduces the Firebase Authentication Emulator.